feat: move today pray to prayfacade
